cleanup, fix pagination
