--- src/views/business/client/orders/OrdersManagement.api.ts
@@ -1,7 +1,0 @@
-export enum Api {
-  getClient = '/userClient/getClient',
-  getShops = '/shippingInvoice/shopsByClient',
-  getOrders = '/business/platformOrder/ordersByShop',
-  postOrders = '/business/platformOrder/orderManagement',
-  getRecipientInfo = '/business/platformOrder/recipientInfo',
-}
